Fixed bugs, restructured stuff :-)
~ Orphans

Changed around how orphans are handled. Previously I had a callback that I gave
to the Tracer that it used to check if orphans had been read. This feels kinda
messy for various reasons and instead the reaper_thread now just calls a public
function of Tracer (notify_orphan) and the tracer handles everything.

~ diagnose_wait_status

Split the code from diagnose_status in forktrace.cpp into a separate global
helper function that takes a wait status and generates a string describing it.
This is generally useful for generating useful debugging messages (see below).

~ BadTraceError

Reverted previous modifications to this class. Instead that error functionality
is done by the diagnose_bad_event function which builds a BadTraceError object.

~ diagnose_bad_event

A helper method that builds a BadTraceError object using a wait status. Should
be called when a wait status is not what was expected. The function probes
around the tracee and adds a bunch of information to the message to make it
more useful (including using diagnose_wait_status).

~ Multiple leaders

The Tracee class can now handle multiple leaders. Why not? There are a few
things I'll need to do in the future. For now the Leader struct only has one
field. This is preferred over having just a map of raw booleans since it keeps
the code more readable and I probably plan on adding more fields later anyway.

~ Tracee::remove_tracee

This hack had to be there previously so that the WaitCall class was able to
remove tracees from the tracer when a wait succeeded. I changed this so that
the WaitCall class directly does this on the tracee (so now the WaitCall class
has to be a friend, not Tracee). This removes the middle man. As a result, the
Tracer::find() function is no longer needed.

~ Tracee enum

Instead of having a bunch of booleans for 'dead', 'stopped', etc. I just made
an enum for it. Not an enum class cos I don't wanna be typing Tracee::Status::
all the time - just typing Tracee:: is enough for me :-(.

~ All public member functions of Tracer now use locking and are thread safe.

~ Minor refactoring

Split code out from handle_wait_notification to new function handle_stopped.

Split code out from forktrace to new function run.

~ Tracer::step() et al

Made a bunch of changes here to ensure things like deadlocks couldn't happen
(e.g., by accidentally waiting when all tracees were already stopped). I also
forgot to update tracee.stopped (now tracee.state) when resuming all tracees,
so that is fixed.

The code actually seems to work now but I'm not entirely sure what I changed
that fixed it (I should do a diff of the two commits later to check this).
The code can run the full example without locking up or crashing at all.

Yeet this was a long commit message... This was fun