Now draws diagrams + other big changes
Changed colour to use Colour enum instead of fmt::text_style directly. There
is some justification in the comments of terminal.hpp. Moved the colour code
from log.cpp to terminal.cpp. Still need to try to get curses colours to be
the same as the libfmt console colours.

Copied over (and modified slightly for style and whatnot) the old diagram
rendering code. There's still some issues with drawing lines for signal sends
that I'll need to have a look at later.

Copied over the old ScrollView code too (modified slightly for style).

Added a bunch of new commands for interactive mode (similar to old version)
to start/stop/resume the tracees from the console as well as view lists of
active tracees and print out some debug info.

Added commands to draw the diagram and to access scrollview (like before). The
new scrollview command doesn't support restarting or stepping from inside the
scrollview yet (will do that later...).

Added commands and command-line options to set the diagram rendering config
(e.g. which events to show and the lane width etc.)

Fixed infinite loop that arises when run without a reaper (since the program
keeps waiting for all orphans to be collected via notify_orphan()). Did this
by adding an extra check into a new helper function that does what I used to
do by just doing while (!step()) { }. Support for --no-reaper isn't essential
since it's more of a debug option but it definitely helps.

Had to change Log::ERR to Log::ERROR to avoid clash with curses #define ERR.

start_tracee now unblocks all signals in the child so that the child doesn't
inherit the tracer's blocked signals (which would interfere with the test).

Did some shuffling around so that the BlockingCall class no longer needs to
be in the tracer.hpp header. Can't be bothered to do the same for the Tracee
struct though (I'd have to add an extra level of indirection for that to work).

Added code in tracer.cpp to handle signal send events via kill/tkill/tgkill.
The code in process.hpp/.cpp was already there to handle this. Fixed potential
issue in notify_signal_send with ordering of the death event and KillEvent.

Also rearranged the functions in tracer.cpp (so the git diff will make it look
like a lot has been changed when it's actually not that much).