Displays scroll-view automatically (instant mode)
If the diagram is too big to display directly to the terminal when in instant
mode, we now just automatically choose the scroll-view for them (previously
they would have had to re-run in interactive mode and choose the scroll view
manually there).

There's also a new command-line option that can be used to force forktrace to
always use the scroll-view when in instant mode, even if the diagram can fit.