ScrollView
Fixed issue with screen size that was being drawn on to (there was a missing
row at the bottom where nothing was drawn).

Have noticed an issue with resizing the ScrollView which causes mayhem in some
circumstances (however it seems to resize completely fine in other cases). Will
need to investigate this further.