Changed way exceptions were thrown
When intercepting certain SystemErrors, I now rethrow the exception with a raw
use of "throw;" rather than "throw e;". The former usage will rethrow whatever
the current exception is, while the latter usage creates a copy and throws that.
Shouldn't affect the behaviour of the program except to minimise unneeded copying.